requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pyyaml>=6.0
joblib>=1.3
tqdm>=4.66
